Fixes 3196 and forces theta to be a eigen vector of doubles in laplace_marginal_density_est
